prepare TokenInput component for export
